Reduce the number of delete operations happening on the sec_audit table
[#141988569] https://www.pivotaltracker.com/story/show/141988569

+++ server/src/main/java/org/cloudfoundry/identity/uaa/audit/JdbcFailedLoginCountingAuditService.java
@@ -12,20 +12,25 @@
  *******************************************************************************/
 package org.cloudfoundry.identity.uaa.audit;
 
+import org.cloudfoundry.identity.uaa.util.TimeService;
+import org.cloudfoundry.identity.uaa.util.TimeServiceImpl;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import java.sql.Timestamp;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * An audit service that subscribes to audit events but only saves enough data
  * to answer queries about consecutive
  * failed logins.
- *
- * @author Dave Syer
  */
 public class JdbcFailedLoginCountingAuditService extends JdbcAuditService {
 
     private int saveDataPeriodMillis = 24 * 3600 * 1000; // 24hr
+    private long timeBetweenDeleteMillis = 1000*30;
+
+    private AtomicLong lastDelete = new AtomicLong(0);
+    private TimeService timeService = new TimeServiceImpl();
 
     public JdbcFailedLoginCountingAuditService(JdbcTemplate template) {
         super(template);
@@ -38,6 +43,10 @@ public class JdbcFailedLoginCountingAuditService extends JdbcAuditService {
         this.saveDataPeriodMillis = saveDataPeriodMillis;
     }
 
+    public void setTimeService(TimeService timeService) {
+        this.timeService = timeService;
+    }
+
     @Override
     public void log(AuditEvent auditEvent) {
         switch (auditEvent.getType()) {
@@ -47,13 +56,21 @@ public class JdbcFailedLoginCountingAuditService extends JdbcAuditService {
                 getJdbcTemplate().update("delete from sec_audit where principal_id=?", auditEvent.getPrincipalId());
                 break;
             case UserAuthenticationFailure:
-                getJdbcTemplate().update("delete from sec_audit where created < ?",
-                                new Timestamp(System.currentTimeMillis()
-                                                - saveDataPeriodMillis));
+                periodicDelete();
                 super.log(auditEvent);
                 break;
             default:
                 break;
+        }
+    }
+
+    protected void periodicDelete() {
+        long now = timeService.getCurrentTimeMillis();
+        long lastCheck = lastDelete.get();
+        if (now - lastCheck > timeBetweenDeleteMillis && lastDelete.compareAndSet(lastCheck, now)) {
+            getJdbcTemplate().update("delete from sec_audit where created < ?",
+                                     new Timestamp(System.currentTimeMillis()
+                                                       - saveDataPeriodMillis));
         }
     }
